requirements.txt
torchgpipe
fvcore==0.1.2.post20210128
fire==0.3.1
opencv-python==4.4.0.46
numpy==1.19.5
requests==2.25.1
matplotlib==3.3.4